Move module/discovery/import driver out-of-line into MLIRGenModule.cpp (review doc 1)
Third slice of the TU split: the module driver family (report,
loadMainSourceFile/loadSourceFile/loadSourceBuf, mlirGenSourceFile,
mlirDiscoverAllDependencies, mlirCodeGenModule, namespace registration,
include loading, shared-lib import; 24 methods, ~930 lines) moves from
inline bodies in MLIRGenImpl.h to MLIRGenModule.cpp.

Method-level #ifdef guards (GENERATE_IMPORT_INFO_USING_D_TS_FILE) are
kept around both the declarations and the moved definitions.

Verified: MLIRGenTests 12/12, full ctest 681/681 (JIT + AOT).
